# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/book_count/click": {
            "post": {
                "tags": [
                    ""
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer {{Authorization}}"
                        }
                    }
                ],
                "requestBody": {
                    "description": "請求body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id",
                                    "book_type_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "description": "",
                                        "type": "integer"
                                    },
                                    "book_type_id": {
                                        "description": "",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "返回響應資料",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "errcode",
                                        "timestamp",
                                        "data"
                                    ],
                                    "properties": {
                                        "errcode": {
                                            "description": "錯誤碼",
                                            "type": "integer"
                                        },
                                        "timestamp": {
                                            "description": "",
                                            "type": "integer"
                                        },
                                        "data": {
                                            "description": "返回資料",
                                            "required": [
                                                "msg"
                                            ],
                                            "properties": {
                                                "msg": {
                                                    "description": "",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}